requirements.txt
numpy==1.17.2
pandas==0.25.3
scikit-learn==0.22.2.post1
scipy==1.4.1
tensorboardX==1.8
-f https://download.pytorch.org/whl/torch_stable.html
torch==1.7.1+cu110
torchaudio==0.7.2
torchvision==0.8.2+cu110
tqdm==4.32.1
transformers==4.18.0
boto3==1.10.27
sentencepiece==0.1.83
mysql-connector-python==8.0.19
pyarrow==6.0.1
